requirements.txt
requests== 2.22.0
python-dotenv==0.10.3
asciitable==0.8.0
